// samples/swr-with-effect/src/gen/endpoints/pets/pets.effect.ts
export const ListPetsResponseItem = S.extend(
  S.Union(
    S.extend(
      S.Union(
        S.Struct({
          cuteness: S.Number,
          breed: S.Literal('Labradoodle'),
        }),
        S.Struct({
          length: S.Number,
          breed: S.Literal('Dachshund'),
        }),
      ),
      S.Struct({
        barksPerMinute: S.optional(S.Number),
        type: S.Literal('dog'),
      }),
    ),
    S.Struct({
      petsRequested: S.optional(S.Number),
      type: S.Literal('cat'),
    }),
  ),
  S.Struct({
    '@id': S.optional(S.String),
    id: S.Number,
    name: S.String,
    tag: S.optional(S.String),
    email: S.optional(S.String.pipe(S.pattern(/^[^\s@]+@[^\s@]+\.[^\s@]+$/))),
    callingCode: S.optional(S.Literal('+33', '+420')),
    country: S.optional(S.Literal("People's Republic of China", 'Uruguay')),
  }),
);
export const ListPetsResponse = S.Array(ListPetsResponseItem);

/**
 * @summary Create a pet
 */
export const CreatePetsBodyItem = S.Struct({
  name: S.String,
  tag: S.String,
});